+++ ipd/ppp/__init__.py
@@ -1,4 +1,6 @@
 from ipd.dev import LazyModule
+
 server = LazyModule('ipd.ppp.server')
+from ipd.ppp.specmodels import *
 from ipd.ppp.client import *
 from ipd.ppp.defaults import *
