+++ ipd/ppp/alembic/versions/51efb10efe65_initial_setup.py
@@ -1,0 +1,30 @@
+"""Initial setup
+
+Revision ID: 51efb10efe65
+Revises: 
+Create Date: 2024-10-07 12:42:02.976923
+
+"""
+from typing import Sequence, Union
+import sqlmodel
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '51efb10efe65'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
