+++ ipd/ppp/server/ppp_server.spec
@@ -18,7 +18,8 @@ IncludeCmd: yes
     git clone https://github.com/willsheffler/willutil /ppp/lib/willutil
     git clone https://github.com/willsheffler/wills_pymol_crap /ppp/lib/wills_pymol_crap
     bash /opt/Miniforge3-Linux-x86_64.sh -b -p /opt/mamba
-    /opt/mamba/bin/mamba install -q -y -c schrodinger pymol-bundle sqlmodel fastapi[standard] uvicorn[standard] ordered-set pyyaml "pip<25" icecream RestrictedPython psycopg2 mysql-connector-python assertpy uvloop
+    /opt/mamba/bin/mamba install -q -y -c schrodinger pymol-bundle sqlmodel fastapi[standard] uvicorn[standard] ordered-set pyyaml "pip<25" icecream RestrictedPython psycopg2 mysql-connector-python assertpy uvloop alembic alembic-utils
+    /opt/mamba/bin/pip install alembic-postgresql-enum
     for lib in ipd willutil wills_pymol_crap; do
         /opt/mamba/bin/pip install -e /ppp/lib/$lib
         git config --global --add safe.directory /ppp/lib/$lib
